requirements.txt
# Usage: pip install -r requirements.txt

# Core AI - API requirements
flask==2.1.2
Pillow==9.1.1

# YOLOv5 requirements
# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
protobuf<=3.20.1

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
ipython
psutil
thop
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow

# Timm requirements
# Model index --------------------------------
model-index==0.1.10
jinja2==3.0

# Docs ----------------------------------------
mkdocs
mkdocs-material
mdx_truly_sane_lists
mkdocs-awesome-pages-plugin
